DEV-58: move deleteDupVersionsFromCourse function to dbCleanUps.js
